--- .ruby-version
@@ -1,1 +1,0 @@
-3.0.2
