--- bin/deploy
@@ -1,32 +1,0 @@
-#!/usr/bin/env ruby
-$stdout.sync = $stderr.sync = true
-require 'bundler/setup'
-require 'dotenv/load'
-require_relative '../website'
-
-name = ARGV.shift
-if name.nil?
-  puts "Usage: deploy {name} [--force]"
-  puts "Where name is cloudamqp or cloudmqtt etc."
-  exit 1
-end
-force = ARGV.shift == '--force'
-puts "Deploying #{name} #{force ? 'with full S3 refresh' : ''}"
-repo_url = if ENV['RACK_ENV'] == 'development'
-             "file://#{ENV['REPO_URL']}/#{name}-website"
-           else
-             auth = "#{ENV.fetch 'OAUTH_TOKEN'}:x-oauth-basic"
-             "https://#{auth}@github.com/84codes/#{name}-website.git"
-           end
-domain = "www.#{name}.com"
-path = "/tmp/#{name}-#{rand(1000..9999)}"
-begin
-  FileUtils.mkdir_p path
-  Dir.chdir path do
-    system "git clone --depth 1 #{repo_url} ."
-    Website.new(domain).upload(force_deploy: force)
-  end
-  puts "#{domain} deployed"
-ensure
-  FileUtils.rm_rf path
-end
