+++ exe/deploy
@@ -1,0 +1,15 @@
+#!/usr/bin/env ruby
+
+require_relative '../lib/website/deployer'
+
+name = ARGV.shift
+if name.nil?
+  puts "Usage: deploy {name} [--force]"
+  puts "Where name is cloudamqp or cloudmqtt etc."
+  exit 1
+end
+force = ARGV.shift == '--force'
+puts "Deploying #{name} #{force ? 'with full S3 refresh' : ''}"
+domain = "www.#{name}.com"
+Website::Deployer.new(domain).upload(force_deploy: force)
+puts "#{domain} deployed"
