+++ lib/website/deployer/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Website
+  module Deployer
+    VERSION = "0.1.0"
+  end
+end
